pyproject.toml
[tool.poetry]
name = "astrotoolz-api"
version = "0.1.0"
description = ""
authors = ["zsoltcs <cszs1123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
pyyaml = "^6.0.1"
colorlog = "^6.8.2"
pyhumps = "^3.8.0"
astrotoolz = {path = "C:/Users/Zsolt/data/dev/astrotoolz"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
